MF-003 Extract Calculation Result to new file
