+++ src/components/Calculator.tsx
@@ -7,15 +7,14 @@ import {
   FormHelperText,
 } from "@material-ui/core";
 import { FareService, TravelRoute } from "../services/fare.service";
-import { getNameFromStation } from "../services/util.service";
 import {
   METRO_STATION,
   BTS_SILOM_STATION,
-  LineType
 } from "../types";
 import { STATION_NAME, STATION_NAME_KEY, Station } from "../data/StationName";
-import "../styles/Calculator.scss";
 import { FindRouteMethod } from "../services/graph.service";
+import CalculationResult from "./CalculationResult";
+import "../styles/Calculator.scss";
 
 const Calculator = () => {
   const [source, setSource] = useState<string>("");
@@ -179,63 +178,4 @@ const SelectStationComponent = ({
   );
 };
 
-function CalculationResult(travelRoute: TravelRoute): React.ReactNode {
-  return (
-    <div>
-      <div>Fare: {travelRoute.fare}</div>
-      <div className="travel-route-container">
-        {travelRoute.route.map((routeSegment, segmentIndex) => {
-          let interchangeDottedLine = null;
-          if (segmentIndex > 0) {
-            interchangeDottedLine = getInterChangeLine(routeSegment.lineType, travelRoute.route[segmentIndex - 1].lineType);
-          }
-          const route = routeSegment.route.map((stationKey, index) => {
-            const dottedLine = getDottedLine(routeSegment.lineType);
-            const stationIcon = getStationIcon(routeSegment.lineType);
-            const stationName = getNameFromStation(stationKey);
-            return (
-              <section key={stationKey}>
-                {index > 0 && dottedLine}
-                <section className="station-container">
-                  {stationIcon}
-                  <div>{stationName}</div>
-                </section>
-              </section>
-            );
-          });
-
-          return (
-            <section key={routeSegment.lineType + "-" + segmentIndex}>
-              {interchangeDottedLine}
-              {route}
-            </section>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
 export default Calculator;
-
-const getDottedLine = (currentLineType: LineType) => {
-  const dottedLineStyle = currentLineType === LineType.MRT_BLUE
-    ? "mrt-blue-dotted-line"
-    : "bts-silom-dotted-line";
-  return <div className={dottedLineStyle}></div>;
-}
-
-const getStationIcon = (currentLineType: LineType) => {
-  const iconStyle = currentLineType === LineType.MRT_BLUE
-    ? "mrt-blue-icon"
-    : "bts-silom-icon";
-  return <div className={iconStyle}></div>;
-}
-
-const getInterChangeLine = (currentLineType: LineType, prevLineType: LineType) => {
-  if (currentLineType === prevLineType) {
-    return getDottedLine(currentLineType);
-  } else {
-    return <div className="interchange-dotted-line"></div>;
-  }
-}
