+++ src/styles/Calculator.scss
@@ -25,52 +25,3 @@
 .w-100 {
   width: 100%;
 }
-
-.station-container {
-  display: flex;
-  padding: 5px;
-}
-
-.mrt-blue-icon {
-  width: 16px;
-  height: 16px;
-  background-color: blue;
-  margin-right: 5px;
-  border-radius: 9px;
-  border: 1px solid black;
-}
-
-.mrt-blue-dotted-line {
-  width: 0px;
-  height: 20px;
-  margin-left: 11px;
-  border-left: 5px dotted blue;
-}
-
-.bts-silom-icon {
-  width: 16px;
-  height: 16px;
-  background-color: #00FF00;
-  margin-right: 5px;
-  border-radius: 9px;
-  border: 1px solid black;
-}
-
-.bts-silom-dotted-line {
-  width: 0px;
-  height: 20px;
-  margin-left: 11px;
-  border-left: 5px dotted #00FF00;
-}
-
-.interchange-dotted-line {
-  width: 0px;
-  height: 20px;
-  margin-left: 11px;
-  border-left: 5px dotted #AAAAAA;
-}
-
-.travel-route-container {
-  overflow-y: auto;
-  max-height: 600px;
-}
